Split TRACK_REQUESTS into two report modules
and introduce `ReportBaseTrackRequests` as a common superclass, so that we can put most of the common logic between "download selected" and "download all" there.
