+++ lib/Constants.js
@@ -603,9 +603,16 @@ ReportType.init({
   TRACK_REQUESTS: {
     disabled: false,
     formats: [ReportFormat.CSV],
-    label: 'Track Requests',
+    label: 'Track Requests: Download All',
     scope: [AuthScope.STUDY_REQUESTS],
     suffix: 'TrackRequests',
+  },
+  TRACK_REQUESTS_SELECTED: {
+    disabled: false,
+    formats: [ReportFormat.CSV],
+    label: 'Track Requests: Download Selected',
+    scope: [AuthScope.STUDY_REQUESTS],
+    suffix: 'TrackRequestsSelected',
   },
   WARRANT_TRAFFIC_SIGNAL_CONTROL: {
     disabled: false,
