+++ lib/io/storage/StoragePath.js
@@ -311,7 +311,7 @@ class StoragePath {
    */
   static async forReport(report) {
     const { type } = report;
-    if (type === ReportType.TRACK_REQUESTS) {
+    if (type === ReportType.TRACK_REQUESTS || type === ReportType.TRACK_REQUESTS_SELECTED) {
       return StoragePath.forStudyRequestReport(report);
     }
     if (type.reportExportMode === ReportExportMode.COLLISIONS) {
