+++ lib/reports/ReportFactory.js
@@ -11,6 +11,7 @@ import ReportIntersectionSummary from '@/lib/reports/ReportIntersectionSummary';
 import ReportPeakHourFactor from '@/lib/reports/ReportPeakHourFactor';
 import ReportSpeedPercentile from '@/lib/reports/ReportSpeedPercentile';
 import ReportTrackRequests from '@/lib/reports/ReportTrackRequests';
+import ReportTrackRequestsSelected from '@/lib/reports/ReportTrackRequestsSelected';
 import ReportWarrantTrafficSignalControl from '@/lib/reports/ReportWarrantTrafficSignalControl';
 
 /**
@@ -65,6 +66,7 @@ ReportFactory.registerInstance(new ReportIntersectionSummary());
 ReportFactory.registerInstance(new ReportPeakHourFactor());
 ReportFactory.registerInstance(new ReportSpeedPercentile());
 ReportFactory.registerInstance(new ReportTrackRequests());
+ReportFactory.registerInstance(new ReportTrackRequestsSelected());
 ReportFactory.registerInstance(new ReportWarrantTrafficSignalControl());
 
 export default ReportFactory;
