+++ lib/reports/ReportIdParser.js
@@ -115,51 +115,16 @@ async function parseStudyReportId(type, rawId) {
   };
 }
 
-async function parseStudyRequestReportId(type, rawId) {
-  const parts = rawId.split('/');
-  if (parts.length !== 2) {
-    throw new InvalidReportIdError(rawId);
-  }
-
-  const idType = parts.shift();
-  if (idType === 'ids') {
-    const studyRequestIdsStr = parts.shift();
-    const ids = [];
-    const idStrs = studyRequestIdsStr.split(',');
-    idStrs.forEach((idStr) => {
-      const id = parseInt(idStr, 10);
-      if (Number.isNaN(idStr)) {
-        throw new InvalidReportIdError(rawId);
-      }
-      ids.push(id);
-    });
-    return {
-      ids,
-      uuid: null,
-    };
-  }
-  if (idType === 'uuid') {
-    const uuid = parts.shift();
-    return {
-      ids: null,
-      uuid,
-    };
-  }
-  throw new InvalidReportIdError(rawId);
-}
-
 /**
  * @namespace
  */
 const ReportIdParser = {
   parseCollisionReportId,
   parseStudyReportId,
-  parseStudyRequestReportId,
 };
 
 export {
   ReportIdParser as default,
   parseCollisionReportId,
   parseStudyReportId,
-  parseStudyRequestReportId,
 };
