+++ tests/jest/unit/reports/ReportIdParser.spec.js
@@ -10,7 +10,6 @@ import { InvalidReportIdError } from '@/lib/error/MoveErrors';
 import {
   parseCollisionReportId,
   parseStudyReportId,
-  parseStudyRequestReportId,
 } from '@/lib/reports/ReportIdParser';
 
 jest.mock('@/lib/db/CentrelineDAO');
@@ -130,32 +129,5 @@ test('ReportIdParser.parseStudyReportId [neither speed- nor TMC-related]', async
     categoryId: 1,
     countGroupId: 5678,
     study: resolvedValue,
-  });
-});
-
-test('ReportIdParser.parseStudyRequestReportId [invalid]', async () => {
-  expect(parseStudyRequestReportId('')).rejects.toBeInstanceOf(InvalidReportIdError);
-  expect(parseStudyRequestReportId('/')).rejects.toBeInstanceOf(InvalidReportIdError);
-  expect(parseStudyRequestReportId('ids')).rejects.toBeInstanceOf(InvalidReportIdError);
-  expect(parseStudyRequestReportId('ids/')).rejects.toBeInstanceOf(InvalidReportIdError);
-  expect(parseStudyRequestReportId('ids/a,b,c')).rejects.toBeInstanceOf(InvalidReportIdError);
-  expect(parseStudyRequestReportId('foo/bar')).rejects.toBeInstanceOf(InvalidReportIdError);
-});
-
-test('ReportIdParser.parseStudyRequestReportId [ids]', async () => {
-  expect(parseStudyRequestReportId('ids/42')).resolves.toEqual({
-    ids: [42],
-    uuid: null,
-  });
-  expect(parseStudyRequestReportId('ids/6,17,73')).resolves.toEqual({
-    ids: [6, 17, 73],
-    uuid: null,
-  });
-});
-
-test('ReportIdParser.parseStudyRequestReportId [uuid]', async () => {
-  expect(parseStudyRequestReportId('uuid/aa73e91b-6aaf-43c7-a368-19a2c81e527e')).resolves.toEqual({
-    ids: null,
-    uuid: 'aa73e91b-6aaf-43c7-a368-19a2c81e527e',
   });
 });
